Move repeated logoc to the NotificationsIcon component
