--- src/components/ApplicantNotificationIcon/index.ts
@@ -1,1 +1,0 @@
-export { ApplicantNotificationIconContainer as ApplicantNotificationIcon } from "./container";
