+++ src/components/ApplicantNotificationIcon/index.tsx
@@ -1,0 +1,7 @@
+import React, { FunctionComponent } from "react";
+import { useHasUnreadApplicantNotifications } from "$hooks";
+import { NotificationsIcon } from "$components/NotificationsIcon";
+
+export const ApplicantNotificationIcon: FunctionComponent = () => {
+  return <NotificationsIcon useHasUnreadNotifications={useHasUnreadApplicantNotifications} />;
+};
