--- src/components/CompanyNotificationsIcon/index.ts
@@ -1,1 +1,0 @@
-export { CompanyNotificationsIconContainer as CompanyNotificationsIcon } from "./container";
