+++ src/components/CompanyNotificationsIcon/index.tsx
@@ -1,0 +1,7 @@
+import React, { FunctionComponent } from "react";
+import { useHasUnreadCompanyNotifications } from "$hooks";
+import { NotificationsIcon } from "$components/NotificationsIcon";
+
+export const CompanyNotificationsIcon: FunctionComponent = () => {
+  return <NotificationsIcon useHasUnreadNotifications={useHasUnreadCompanyNotifications} />;
+};
